feat(trigger): Smoothing and median ratio threshold determination
Fairly slap-dash implementation of a simple gaussian smoothing to the detect coa
output, and addition of a "median ratio" dynamic threshold method.

Also reverts the change to always using "COA" to determine the trigger time -
this is correct as locate() only uses the "COA" stream to determine the
location / origin time.